requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
aiosqlite==0.22.1
anyio==4.13.0
asarPy==1.0.1
attrs==26.1.0
audioop-lts==0.2.2
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
cryptography==46.0.7
discord.py==2.7.1
Flask==3.1.3
frozenlist==1.8.0
google-api-core==2.30.3
google-api-python-client==2.194.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
google-auth-oauthlib==1.3.1
googleapis-common-protos==1.74.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
multidict==6.7.1
oauthlib==3.3.1
pillow==12.2.0
propcache==0.4.1
proto-plus==1.27.2
protobuf==7.34.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pylast==7.0.2
pyparsing==3.3.2
python-dotenv==1.2.2
pytz==2026.1.post1
redis==7.4.0
requests==2.33.1
requests-oauthlib==2.0.0
spotipy==2.26.0
uritemplate==4.2.0
urllib3==2.6.3
Werkzeug==3.1.8
yarl==1.23.0